refactor: 테스트 서포터 클래스 파일 경로 수정 (#18)
