+++ inpeak/src/test/java/com/blooming/inpeak/common/config/SecurityConfigTest.java
@@ -3,7 +3,7 @@ package com.blooming.inpeak.common.config;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.blooming.inpeak.IntegrationTestSupport;
+import com.blooming.inpeak.support.IntegrationTestSupport;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
