+++ inpeak/src/test/java/com/blooming/inpeak/support/IntegrationTestSupport.java
@@ -1,5 +1,6 @@
-package com.blooming.inpeak;
+package com.blooming.inpeak.support;
 
+import com.blooming.inpeak.TestRedisConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
